Enables instantaneous query from usgs
The issue here is that the USGS plugin requires and
aggregation_durations of 'NONE' but the timeseries plugin does not
support a temporal_resolution value of 'NON' (only SECOND, MINUTE, HOUR,
DAY)

This commit will set query.aggregation_duration to 'NONE' if MINUTE.
SECOND or HOUR are passed in.  This may not be the right solution
for all future plugins.